Create Balancing of Symbols.c
The stack data structure in this program is implemented by dynamic arrays. Each time the stack is full and a push function is called the size of the array doubles. (,{,[ are pushed in to the stack when encounter by the program. ),},] When closing brackets are present in the expression; pop function is called. This function returns the character and a check function checks if it matches the closing brackets as the stacks only contains opening brackets.