Create HackerRank Small Triangles, Large Triangles.c
This programs uses insertion sort for arranging areas in ascending order. Once that is done a comparison is made with the old(non-sorted) area array for finding the positions.  Know the only task left is to arrange the structure array tr in the order of position array. This procedure is achieved by creating a new structure array and placing the values as per position array and finally again putting back the values to tr array.